// dbaas/models/ChangeControlPanel200Response.ts
/**
 * DBaaS
 * Manage your databases using our API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ChangeControlPanel200Response {
    'controlPanelID'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "controlPanelID",
            "baseName": "controlPanelID",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ChangeControlPanel200Response.attributeTypeMap;
    }

    public constructor() {
    }
}