// dbaas/models/CreateDatabase200Response.ts
/**
 * DBaaS
 * Manage your databases using our API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateDatabase200Response {
    'databaseID'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "databaseID",
            "baseName": "databaseID",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateDatabase200Response.attributeTypeMap;
    }

    public constructor() {
    }
}