// dbaas/models/DBDetailsDatabaseMetaData.ts
/**
 * DBaaS
 * Manage your databases using our API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class DBDetailsDatabaseMetaData {
    /**
    * Indicates if the database uses a standalone replica set
    */
    'standaloneReplicaSet'?: boolean;
    /**
    * Indicates if the database is on a private network
    */
    'privateNetwork'?: boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "standaloneReplicaSet",
            "baseName": "standaloneReplicaSet",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "privateNetwork",
            "baseName": "privateNetwork",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DBDetailsDatabaseMetaData.attributeTypeMap;
    }

    public constructor() {
    }
}