// dbaas/models/DBsDetails.ts
/**
 * DBaaS
 * Manage your databases using our API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DBsDetailsDatabasesInner } from '../models/DBsDetailsDatabasesInner';
import { HttpFile } from '../http/http';

export class DBsDetails {
    'databases'?: Array<DBsDetailsDatabasesInner>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "databases",
            "baseName": "databases",
            "type": "Array<DBsDetailsDatabasesInner>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DBsDetails.attributeTypeMap;
    }

    public constructor() {
    }
}