// dbaas/models/GetDatabaseSummaryReports200Response.ts
/**
 * DBaaS
 * Manage your databases using our API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetDatabaseSummaryReports200ResponseCpuUsageInner } from '../models/GetDatabaseSummaryReports200ResponseCpuUsageInner';
import { GetDatabaseSummaryReports200ResponseDisksUsageInner } from '../models/GetDatabaseSummaryReports200ResponseDisksUsageInner';
import { HttpFile } from '../http/http';

export class GetDatabaseSummaryReports200Response {
    'cpuUsage'?: Array<GetDatabaseSummaryReports200ResponseCpuUsageInner>;
    'memoryUsage'?: Array<GetDatabaseSummaryReports200ResponseCpuUsageInner>;
    'disksUsage'?: Array<GetDatabaseSummaryReports200ResponseDisksUsageInner>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "cpuUsage",
            "baseName": "cpuUsage",
            "type": "Array<GetDatabaseSummaryReports200ResponseCpuUsageInner>",
            "format": ""
        },
        {
            "name": "memoryUsage",
            "baseName": "memoryUsage",
            "type": "Array<GetDatabaseSummaryReports200ResponseCpuUsageInner>",
            "format": ""
        },
        {
            "name": "disksUsage",
            "baseName": "disksUsage",
            "type": "Array<GetDatabaseSummaryReports200ResponseDisksUsageInner>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetDatabaseSummaryReports200Response.attributeTypeMap;
    }

    public constructor() {
    }
}