// dbaas/models/GetDatabaseSummaryReports200ResponseDisksUsageInner.ts
/**
 * DBaaS
 * Manage your databases using our API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class GetDatabaseSummaryReports200ResponseDisksUsageInner {
    'size'?: string;
    'usage'?: string;
    'reportedAt'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "size",
            "baseName": "size",
            "type": "string",
            "format": ""
        },
        {
            "name": "usage",
            "baseName": "usage",
            "type": "string",
            "format": ""
        },
        {
            "name": "reportedAt",
            "baseName": "reportedAt",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetDatabaseSummaryReports200ResponseDisksUsageInner.attributeTypeMap;
    }

    public constructor() {
    }
}