// dbaas/models/GetListBackups200Response.ts
/**
 * DBaaS
 * Manage your databases using our API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetListBackups200ResponseBackupsInner } from '../models/GetListBackups200ResponseBackupsInner';
import { HttpFile } from '../http/http';

export class GetListBackups200Response {
    'backups'?: Array<GetListBackups200ResponseBackupsInner>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "backups",
            "baseName": "backups",
            "type": "Array<GetListBackups200ResponseBackupsInner>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetListBackups200Response.attributeTypeMap;
    }

    public constructor() {
    }
}